Resolves compiler warnings with explicit size_t casts
* Adds explicit `(size_t)` casts to memory allocations and API calls in the CLI, driver, and test suite.
* Prevents potential truncation and type mismatch warnings when converting `uint64_t` or `int64_t` values for use with `malloc`, `fwrite`, and library functions.

+++ src/cli/main.c
@@ -1279,7 +1279,7 @@ int main(int argc, char** argv) {
         fm = NULL;
 
         const uint64_t max_c = zxc_compress_bound(in_size);
-        c_dat = malloc(max_c);
+        c_dat = malloc((size_t)max_c);
         if (!c_dat) goto bench_cleanup;
 
 #ifdef _WIN32
